requirements.txt
flask
pandas
scikit-learn==1.2.2
faiss-cpu
joblib
google-cloud-storage
google-cloud-firestore
gunicorn
